+++ src/compat/transformers.test.ts
@@ -1,8 +1,15 @@
 import { assertEquals } from "@std/assert/assert-equals";
 import { assertNotEquals } from "@std/assert/assert-not-equals";
+import { MemoryKvStore } from "../federation/kv.ts";
+import { FederationImpl } from "../federation/middleware.ts";
 import { test } from "../testing/mod.ts";
 import { Follow, Person } from "../vocab/vocab.ts";
 import { actorDehydrator, autoIdAssigner } from "./transformers.ts";
+
+const federation = new FederationImpl<void>({
+  kv: new MemoryKvStore(),
+});
+const context = federation.createContext(new URL("http://example.com/"));
 
 test("autoIdAssigner", async () => {
   const activity = new Follow({
@@ -13,7 +20,7 @@ test("autoIdAssigner", async () => {
       preferredUsername: "bob",
     }),
   });
-  const result = autoIdAssigner(activity);
+  const result = autoIdAssigner(activity, context);
   assertNotEquals(result.id, null);
   assertEquals(
     await result.toJsonLd(),
@@ -43,7 +50,7 @@ test("actorDehydrator()", async () => {
       preferredUsername: "bob",
     }),
   });
-  const result = actorDehydrator(activity);
+  const result = actorDehydrator(activity, context);
   assertEquals(
     await result.toJsonLd(),
     await new Follow({